Added authorization

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -15,6 +15,11 @@ class AttendeeController extends Controller
 
     private $relations = ['user'];
 
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
+
     public function index(Event $event)
     {
         $attendees = $this->loadRelationships(
@@ -45,13 +50,9 @@ class AttendeeController extends Controller
         );
     }
 
-    public function update(Request $request, string $id)
+    public function destroy(Event $event, Attendee $attendee)
     {
-        //
-    }
-
-    public function destroy(string $event, Attendee $attendee)
-    {
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
 
         return response(status: 204);
